refactor: made HomePAger code cleaner
